# openapi:3.0.yml
openapi: '3.0.2'
info:
  title: API Booking
  version: '1.0'
servers:
  - url: http://localhost:3000/api/
paths:
...
  /articles:
    get:
      summary: Récupère tous les articles
      description: Récupère tous les articles enregistrés dans la base de données, ou ceux qui appartiennent à une catégorie donnée si le paramètre `categorieId` est fourni dans la requête.
      parameters:
        - in: query
          name: categorieId
          description: ID de la catégorie des articles à récupérer
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '500':
          description: Une erreur est survenue.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Description de l'erreur survenue.
      tags:
        - Articles

    post:
      summary: Créer un nouvel article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    title: 
                      type: string
                    description: 
                      type: string
                    content: 
                      type: string
                    categorie: 
                      type: number
      responses:
        '201':
          description: Article créé avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - Articles
...
  /bookings:
    get:
      summary: Récupérer toutes les réservations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    room_id:
                      type: integer
                    start_date:
                      type: string
                      format: date-time
                    end_date:
                      type: string
                      format: date-time
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - "Bookings"
    post:
      summary: Créer une réservation
      requestBody:
        description: Données pour créer une réservation
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    email: 
                      type: string
                    phone: 
                      type: string
                    dateCheckIn: 
                      type: string
                    dateCheckOut: 
                      type: string
                    personNumber: 
                      type: number
                    comment:
                      type: string
                    roomId: 
                      type: number
      responses:
        '201':
          description: Réservation créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Non autorisé
        '409':
          description: Conflit
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Erreur serveur
      tags:
        - "Bookings"

# Category Articles
...
components:
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        content:
          type: string
        date_publish:
          type: string
          format: date-time
        date_update:
          type: string
          format: date-time
        images:
          type: array
          items:
            type: number
            $ref: '#/components/schemas/ImageArticle'
        categorie:
          type: number
          $ref: '#/components/schemas/CategorieArticle'
    ImageArticle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
        description:
          type: string
        urlStorage:
          type: string
        article:
          $ref: '#/components/schemas/Article'
          type: number
    CategorieArticle:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Identifiant de la catégorie.
        title:
          type: string
          maxLength: 128
          description: Titre de la catégorie.
        description:
          type: string
          maxLength: 255
          nullable: true
          description: Description de la catégorie.
        articles:
          type: array
          items:
            type: number
            $ref: '#/components/schemas/Article'
          description: Liste des articles de cette catégorie.
...
    Booking:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier for a booking
        uuid:
          type: string
          description: The unique identifier for a booking in UUID format
        phone:
          type: string
          description: The phone number associated with the booking
        email:
          type: string
          description: The email associated with the booking
        dateCheckIn:
          type: string
          format: date-time
          description: The check-in date for the booking
        dateCheckOut:
          type: string
          format: date-time
          description: The check-out date for the booking
        duration:
          type: integer
          description: The duration of the booking in days
        personNumber:
          type: integer
          description: The number of people associated with the booking
        comment:
          type: string
          description: Any comments associated with the booking
        userId:
          type: integer
          description: The user ID associated with the booking
        roomId:
          type: integer
          description: The room ID associated with the booking
        hasPaid:
          type: boolean
          description: Whether the booking has been paid for or not
        price:
          type: number
          description: The price of the booking
        stripePaymentIntent:
          type: string
          description: The Stripe payment intent associated with the booking
        createdOn:
          type: string
          format: date-time
          description: The date and time the booking was created
        updatedOn:
          type: string
          format: date-time
          description: The date and time the booking was last updated
        paymentDate:
          type: string
          format: date-time
          description: The date and time the booking was paid for